# swagger_gen/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentation's GU user service",
        "title": "GU user service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alexander Sharov",
            "url": "http://github.com/kvendingoldo",
            "email": "kvendingoldo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/users": {
            "get": {
                "description": "Get all users on the system",
                "tags": [
                    "user"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/users.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/users.MessageResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new user on the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create new user",
                "parameters": [
                    {
                        "description": "body data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.NewUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/users.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/users.MessageResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "Get users by ID on the system",
                "tags": [
                    "user"
                ],
                "summary": "Get users by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/users.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/users.MessageResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update user on the system",
                "tags": [
                    "user"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User updated info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete user on the system",
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.User": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "type": "string",
                    "example": "39.12355, 27.64538"
                },
                "created_at": {
                    "type": "string",
                    "example": "2021-02-24 20:19:39"
                },
                "id": {
                    "type": "integer",
                    "example": 123
                },
                "name": {
                    "type": "string",
                    "example": "Steven"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-02-24 20:19:39"
                }
            }
        },
        "users.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "users.NewUserRequest": {
            "type": "object",
            "required": [
                "coordinates",
                "name"
            ],
            "properties": {
                "coordinates": {
                    "type": "string",
                    "example": "39.12355, 27.64538"
                },
                "name": {
                    "type": "string",
                    "example": "Bill"
                }
            }
        }
    }
}